Confirm Dialog Box Added

+++ src/app/pages/rpautomation/rpa-action-items/rpa-action-items.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from "@angular/core";
 import { ActivatedRoute, Router } from "@angular/router";
-import { MessageService } from "primeng/api";
+import { ConfirmationService, MessageService } from "primeng/api";
 import { LoaderService } from 'src/app/services/loader/loader.service';
 import { RestApiService } from "../../services/rest-api.service";
 
@@ -27,7 +27,8 @@ export class RpaActionItemsComponent implements OnInit {
     private loader:LoaderService,
     private rest_api:RestApiService,
     private route:ActivatedRoute,
-    private messageService: MessageService,
+    private confirmationService: ConfirmationService,
+    private messageService: MessageService
     ) 
     {
       this.route.queryParams.subscribe((data)=>{
@@ -124,23 +125,36 @@ export class RpaActionItemsComponent implements OnInit {
   deleteAction() {
     this.loader.show();
     let selectedId = this.selectedData[0].id;
-    this.rest_api.deleteActionById(selectedId).subscribe((res:any)=>{
-      this.messageService.add({
-        severity: "success",
-        summary: "Success",
-        detail: "Action Deleted Successfully !!",
-      })
-      this.loader.hide();
-      this.getAllActionItems();
-    },(err) => {
-      this.messageService.add({
-        severity: "error",
-        summary: "Error",
-        detail: "Something Went Wrong !!",
-      })
-      this.loader.hide();
-      this.getAllActionItems();
-    })
+    this.confirmationService.confirm({
+      message: "Are you sure?, You won't be able to revert this!",
+      header: 'Confirmation',
+      icon: 'pi pi-info-circle',
+      accept: () => {
+        this.loader.show();
+        this.rest_api.deleteActionById(selectedId).subscribe((res:any)=>{
+          this.messageService.add({
+            severity: "success",
+            summary: "Success",
+            detail: "Action Deleted Successfully !!",
+          })
+          this.loader.hide();
+          this.getAllActionItems();
+        },(err) => {
+          this.messageService.add({
+            severity: "error",
+            summary: "Error",
+            detail: "Something Went Wrong !!",
+          })
+          this.loader.hide();
+          this.getAllActionItems();
+        })
+      },
+      reject: (type) => {
+        this.loader.hide();
+      },
+      key: "positionDialog"
+    });
+
   }
   
   backToConnection() {
